REST Campus Service Dependency Injection and code refactoring
